The transport provides a serialize mechanism
The HTTP POST serialization logic is provided by cURL or other HTTP
transport libraries. So we need the transport to provide the logic for
serializin the parameters.

However Google Analytics puts restrictions on the size of the POST when
doing batched commits, so we need to be able to handle that outside of
the specific transport implementations.